add config file but there is same warnings in stdout

+++ cub3D.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   cub3D.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: zouaraqa <zouaraqa@student.1337.ma>        +#+  +:+       +#+        */
+/*   By: yaidriss <yaidriss@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/20 06:47:35 by yaidriss          #+#    #+#             */
-/*   Updated: 2023/08/24 16:13:47 by zouaraqa         ###   ########.fr       */
+/*   Updated: 2023/08/26 00:11:00 by yaidriss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -90,8 +90,8 @@ typedef struct s_texture
 }				t_texture;
 typedef struct s_player
 {
-	double	x;
-	double	y;
+	int		x;
+	int		y;
 	double	width;
 	double	height;
 	int		turn_direction;
